requirements.txt
channels==4.0.0
daphne==4.0.0
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.4.0
python-dateutil==2.8.2
